+++ packages/custom_lint/test/mock_fs.dart
@@ -9,10 +9,12 @@ import 'dart:io';
 Future<T> runWithIOOverride<T>(
   FutureOr<T> Function(Stream<String> out, Stream<String> err) testBody, {
   Directory? currentDirectory,
+  bool supportsAnsiEscapes = false,
 }) async {
   final fs = _MockFs(
     stdout,
     currentDirectory ?? Directory.current,
+    supportsAnsiEscapes: supportsAnsiEscapes,
   );
 
   try {
@@ -30,7 +32,10 @@ Future<T> runWithIOOverride<T>(
 }
 
 class _StdoutOverride implements Stdout {
-  _StdoutOverride(this._stdout);
+  _StdoutOverride(
+    this._stdout, {
+    required this.supportsAnsiEscapes,
+  });
 
   final Stdout _stdout;
 
@@ -72,7 +77,7 @@ class _StdoutOverride implements Stdout {
   IOSink get nonBlocking => _stdout.nonBlocking;
 
   @override
-  bool get supportsAnsiEscapes => _stdout.supportsAnsiEscapes;
+  final bool supportsAnsiEscapes;
 
   @override
   int get terminalColumns => _stdout.terminalColumns;
@@ -112,9 +117,12 @@ class _StdoutOverride implements Stdout {
 /// Alternatively, set [IOOverrides.global] to a [_MockFs] instance in your
 /// test's `setUp`, and to `null` in the `tearDown`.
 class _MockFs extends IOOverrides {
-  _MockFs(Stdout out, this._directory)
-      : stdout = _StdoutOverride(out),
-        stderr = _StdoutOverride(out);
+  _MockFs(
+    Stdout out,
+    this._directory, {
+    required bool supportsAnsiEscapes,
+  })  : stdout = _StdoutOverride(out, supportsAnsiEscapes: supportsAnsiEscapes),
+        stderr = _StdoutOverride(out, supportsAnsiEscapes: supportsAnsiEscapes);
 
   @override
   final _StdoutOverride stdout;