+++ packages/custom_lint/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased patch
 
 - Sort lints by severity in the command line (thanks to @kuhnroyal)
+- Fix watch mode not quitting with `q` (thanks to @kuhnroyal)
 
 ## 0.5.3 - 2023-08-29